Support different state animations for different components.
