+++ static/src/components/builder/components/ComponentRender/states/timer.js
@@ -1,0 +1,14 @@
+const TIMER_SECONDS = 5;
+
+export const states = [
+    {
+        taskId: 1,
+        expiration: {
+            get seconds() {
+              return Math.trunc(Date.now() / 1000) + TIMER_SECONDS;
+            }
+        }
+    }
+]
+
+export const intervalSeconds = TIMER_SECONDS + 2;
